fix(solid): update data fetch example
